error solve

+++ C-LANG/pro28.c
@@ -3,16 +3,16 @@
 #include<stdio.h>
 int main()
 {
-    int year,day,month;
+    float year,day,month;
 
     printf("enter no of year: ");
-    scanf("%d",&year);
+    scanf("%f",&year);
 
     printf("enter year into day: ");
-    scanf("%d",&day);
+    scanf("%f",&day);
 
-    day=year*365;
-    month=year*12;
-    printf("%d Years Equal to %d DAYS",year,day);
-    printf("\n%d YEARS Equal to %d MONTHS",year,month);
+    day=year*365.00;
+    month=year*12.00;
+    printf("%.2f Years Equal to %.2f DAYS",year,day);
+    printf("\n%.2f YEARS Equal to %.2f MONTHS",year,month);
 }
